revert old int strides

+++ src/gpuarray_blas_cuda_cublas.c
@@ -878,15 +878,15 @@ static int ddot(
 
 static int hgemv(cb_order order, cb_transpose transA, size_t M, size_t N,
                  float alpha, gpudata *A, size_t offA, size_t lda,
-                 gpudata *X, size_t offX, size_t incX,
-                 float beta, gpudata *Y, size_t offY, size_t incY) {
+                 gpudata *X, size_t offX, int incX,
+                 float beta, gpudata *Y, size_t offY, int incY) {
   return GA_DEVSUP_ERROR;
 }
 
 static int sgemv(cb_order order, cb_transpose transA, size_t M, size_t N,
                  float alpha, gpudata *A, size_t offA, size_t lda,
-                 gpudata *X, size_t offX, size_t incX,
-                 float beta, gpudata *Y, size_t offY, size_t incY) {
+                 gpudata *X, size_t offX, int incX,
+                 float beta, gpudata *Y, size_t offY, int incY) {
   cuda_context *ctx = A->ctx;
   blas_handle *h = (blas_handle *)ctx->blas_handle;
   size_t t;
@@ -940,8 +940,8 @@ static int sgemv(cb_order order, cb_transpose transA, size_t M, size_t N,
 
 static int dgemv(cb_order order, cb_transpose transA, size_t M, size_t N,
                  double alpha, gpudata *A, size_t offA, size_t lda,
-                 gpudata *X, size_t offX, size_t incX,
-                 double beta, gpudata *Y, size_t offY, size_t incY) {
+                 gpudata *X, size_t offX, int incX,
+                 double beta, gpudata *Y, size_t offY, int incY) {
   cuda_context *ctx = A->ctx;
   blas_handle *h = (blas_handle *)ctx->blas_handle;
   size_t t;
@@ -1253,13 +1253,13 @@ static int dgemvBatch(cb_order order, cb_transpose transA,
 
 
 static int hger(cb_order order, size_t M, size_t N, float alpha, gpudata *X,
-                size_t offX, size_t incX, gpudata *Y, size_t offY, size_t incY,
+                size_t offX, int incX, gpudata *Y, size_t offY, int incY,
                 gpudata *A, size_t offA, size_t lda) {
   return GA_DEVSUP_ERROR;
 }
 
 static int sger(cb_order order, size_t M, size_t N, float alpha, gpudata *X,
-                size_t offX, size_t incX, gpudata *Y, size_t offY, size_t incY,
+                size_t offX, int incX, gpudata *Y, size_t offY, int incY,
                 gpudata *A, size_t offA, size_t lda) {
   cuda_context *ctx = X->ctx;
   blas_handle *h = (blas_handle *)ctx->blas_handle;
@@ -1316,7 +1316,7 @@ static int sger(cb_order order, size_t M, size_t N, float alpha, gpudata *X,
 }
 
 static int dger(cb_order order, size_t M, size_t N, double alpha, gpudata *X,
-                size_t offX, size_t incX, gpudata *Y, size_t offY, size_t incY,
+                size_t offX, int incX, gpudata *Y, size_t offY, int incY,
                 gpudata *A, size_t offA, size_t lda) {
   cuda_context *ctx = X->ctx;
   blas_handle *h = (blas_handle *)ctx->blas_handle;
